chore: remove rest of _internals
